requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==41.0.7
Django==4.2.8
django-appconf==1.0.6
django-cryptography==1.1
django-nested-admin==4.0.2 # support nested inlines in admin form
djangorestframework==3.14.0
-e git+https://github.com/ankits16/ManagedStringTestProject.git@f397376faf51f55f285c9c8d28583632a21ea342#egg=lib_managed_string&subdirectory=main/lib_managed_string
pycparser==2.21
python-monkey-business==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
